LayoutTests/fast/css-grid-layout/grid-content-alignment-overflow.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/grid.css" rel="stylesheet">
<script src="../../resources/check-layout.js"></script>
<style>
body {
    margin: 0;
}

.grid {
    -webkit-grid-template-columns: 50px 50px;
    -webkit-grid-template-rows: 100px 100px;
}

.overflowWidth {
    width: 60px;
    height: 300px;
}

.overflowHeight {
    width: 200px;
    height: 150px;
}

.center {
    align-content: center;
    justify-content: center;
}

.centerSafe {
    align-content: center safe;
    justify-content: center safe;
}

.centerUnsafe {
    align-content: center unsafe;
    justify-content: center unsafe;
}

.end {
    align-content: end;
    justify-content: end;
}

.endSafe {
    align-content: end safe;
    justify-content: end safe;
}

.endUnsafe {
    align-content: end unsafe;
    justify-content: end unsafe;
}

</style>
</head>
<body onload="checkLayout('.grid')">

<p>This test checks that the 'overflow' keyword is applied correctly for Content Alignment properties.</p>

<div style="position: relative">
    <div class="grid overflowWidth center" data-expected-width="60" data-expected-height="300">
        <div class="firstRowFirstColumn" data-offset-x="-20" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="-20" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="30" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="30" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowHeight centerUnsafe" data-expected-width="200" data-expected-height="150">
        <div class="firstRowFirstColumn" data-offset-x="50" data-offset-y="-25" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="50" data-offset-y="75" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="100" data-offset-y="-25" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="100" data-offset-y="75" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowWidth centerSafe" data-expected-width="60" data-expected-height="300">
        <div class="firstRowFirstColumn" data-offset-x="0" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="0" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="50" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="50" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowHeight centerSafe" data-expected-width="200" data-expected-height="150">
        <div class="firstRowFirstColumn" data-offset-x="50" data-offset-y="0" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="50" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="100" data-offset-y="0" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="100" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowWidth end" data-expected-width="60" data-expected-height="300">
        <div class="firstRowFirstColumn" data-offset-x="-40" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="-40" data-offset-y="200" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="10" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="10" data-offset-y="200" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowHeight endUnsafe" data-expected-width="200" data-expected-height="150">
        <div class="firstRowFirstColumn" data-offset-x="100" data-offset-y="-50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="100" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="150" data-offset-y="-50" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="150" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowWidth endSafe" data-expected-width="60" data-expected-height="300">
        <div class="firstRowFirstColumn" data-offset-x="0" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="0" data-offset-y="200" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="50" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="50" data-offset-y="200" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <div class="grid overflowHeight endSafe" data-expected-width="200" data-expected-height="150">
        <div class="firstRowFirstColumn" data-offset-x="100" data-offset-y="0" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowFirstColumn" data-offset-x="100" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
        <div class="firstRowSecondColumn" data-offset-x="150" data-offset-y="0" data-expected-width="50" data-expected-height="100"></div>
        <div class="secondRowSecondColumn" data-offset-x="150" data-offset-y="100" data-expected-width="50" data-expected-height="100"></div>
    </div>
</div>

</body>
</html>
