LayoutTests/fast/css-grid-layout/grid-content-alignment-overflow-expected.txt

This test checks that the 'overflow' keyword is applied correctly for Content Alignment properties.

PASS
Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: center and Overflow-Alignment: default

PASS
Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: center and Overflow-Alignment: unsafe

PASS
Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: center and Overflow-Alignment: safe

PASS
Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: center and Overflow-Alignment: safe

PASS
Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: end and Overflow-Alignment: default

PASS
Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: end and Overflow-Alignment: unsafe

PASS
Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: end and Overflow-Alignment: safe

PASS
Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: end and Overflow-Alignment: safe

PASS
Grid container indefinite size using fit-content, hence no possible overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

PASS
Grid container indefinite size using fit-content, hence no possible overflow.
Content-Alignment: center and Overflow-Alignment: unsafe

PASS
Grid container indefinite size using fit-content, but max-size constraints implies that the content-sized tracks overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

PASS
Grid container indefinite size using fit-content, but max-size constraints implies that the content-sized tracks overflow.
Content-Alignment: center and Overflow-Alignment: unsafe

PASS
Grid container indefinite size using fit-content, but min-size constraints implies that the content-sized tracks don't overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

PASS
Grid container indefinite size using fit-content, but min-size constraints implies that the content-sized tracks don't overflow.
Content-Alignment: center and Overflow-Alignment: unsafe
